# Pcvl 860 more clever tensorproduct for noisy state generation