--- a/tests/test_processor.py
+++ b/tests/test_processor.py
@@ -28,13 +28,12 @@
 # SOFTWARE.
 
 import pytest
-import numpy as np
 from unittest.mock import patch
 
 import perceval as pcvl
 from perceval import BSDistribution
 from perceval.components import Circuit, Processor, BS, Source, catalog, UnavailableModeException, Port, PortLocation, \
-    PS, PERM, Detector
+    PERM, Detector
 from perceval.utils import BasicState, StateVector, SVDistribution, Encoding, NoiseModel
 from perceval.backends import Clifford2017Backend
 
@@ -277,7 +276,7 @@ def test_empty_output(mock_warn):
     p.min_detected_photons_filter(2)
     p.with_input(BasicState([0, 1, 0]))
 
-    with LogChecker(mock_warn, expected_log_number=1):  # Normalize is called once
+    with LogChecker(mock_warn, expected_log_number=2):  # Normalize is called once, a void tensor product is called once
         res = p.probs()["results"]
     assert res == BSDistribution()