Some benchmarks on my laptop, with a fully imperfect source, for the function `Processor.with_input`:

To generate a state with 9 photons [1, 1, ...]
Before this PR: 50 seconds
After this PR: 40 seconds

To generate a state with 9 photons in 18 modes [1, 0, 1, 0, ...]
Before this PR: 110 seconds
After this PR: 40 seconds  (the time is now almost independant of empty modes)

To generate a state with 9 photons in the same mode [9]
Before this PR: 160 seconds
After this PR: 38 seconds

The modifications of the tensor product method were also done in the Simulator. Again, the benchmark uses a fully imperfect source, but is now for the function `Processor.probs`, using SLOS.

Using a random unitary, with 7 modes and photons [1, 1... ], no photons filter
Before this PR: 32 seconds
After this PR: 30 seconds
Globally, 6 to 8% speedup using different numbers of photons and modes

Using an empty circuit
Before this PR: 7.9 seconds
After this PR: 8 seconds
Globally, 1 to 2% slowdown

[benchmark_sources.txt](https://github.com/user-attachments/files/18329910/benchmark_sources.txt)
[benchmark sources + probs.txt](https://github.com/user-attachments/files/18329911/benchmark.sources.%2B.probs.txt)